Remove placeholder system
Restructure TOML significantly and require proper ordering of structs in
the file.
Also restructure TOML to make the struct fields a list (because
obviously that order is important), but I didn't notice before that
toml++ will give them to me in alphabetical and not in definition order
(which is fine, since the spec says the order is not guaranteed).