Let window system handle input and pass it to input system instead
Calling sf::Keyboard::isKeyPressed a lot slowed the program down
noticably, so I instead keep track of the current input state via the
events in the event SFML event queue.
Just for future reference (it doesn't mean as much as it looks like, I
know): the FPS went from ~600 to 12000.